Adding e/m counters for different ids
Number of electrons or muons passing a cut (i.e. an ID criteria) is
added first to the event stream and then to the final state and
then to ntuples. Currently only implemented for ZPlusXFinalState, so it
will break things in ZZ.

Also would be a good idea to made the connection between ID and counters
more constrained.