Adding iso to WZ muon IDs to get around error
When ISO is not included in muon IDs and called separately, during the
WZCrossCleaning module the following error arises, which I don't
understand (I think it could be due to the cut string parsing, but it's
        not at all clear that it comes from this):

std::bad_alloc exception
----- Begin Fatal Exception 21-Nov-2016 12:29:24 CST-----------------------
An exception of category 'BadAlloc' occurred while
   [0] Processing run: 1 lumi: 7564 event: 1512277
   [1] Running path 'flowFlowPath'
   [2] Calling event method for module ElectronRegressionValueMapProducer/'electronRegressionValueMapProducer'
   [3] Using EventSetup component CaloGeometryBuilder/'' to make data CaloGeometry/'' in record CaloGeometryRecord
   [4] Using EventSetup component EcalBarrelGeometryFromDBEP/'' to make data CaloSubdetectorGeometry/'EcalBarrel' in record EcalBarrelGeometryRecord
Exception Message:
A std::bad_alloc exception was thrown.
The job has probably exhausted the virtual memory available to the process.
----- End Fatal Exception -------------------------------------------------